refac: Move Binstar class into new client module
